feat(Interactor): ability to modify interactor settings dynamically
The InteractorFacadeSettingsModifier provides a way of specifying
setting modifiers for a collection of interactors and then apply
the settings dynamically.

This can be used for updating the grab action on an interactor to
be something different from the default based on the interactable
being touched. For example, the default grab could be grip press
but on touch of an interactable it could change the interactor
grab action to be trigger press, and then on untouch resort to the
original cached action.