fix(Interactions): allow for multiple grab orientation handles
There was an issue with the interactable when using multiple grab
orientation handles as it was always set up by default to get the
first GameObject relation, which meant when multiple orientation
handles were required then the nested prefab logic would also need
changing.

It's now possible to do all of the logic required in the event flow
due to a new event on the GameObjectRelations component.

The new functionality is to attempt to match the Interactor GameObject
as the key in the GameObjectRelations and if it's not found then the
new event will simply attempt to look up the first relation as this
is what happened previously.