feat(Interactables): apply more specific namespace to interactables
Previously, the Interactables were in a namespace that was at a higher
level of the Interactors whereas Interactables should be at the same
level.

There is already an Interactables namespace used by both interactors
and interactables so the best solution is to just have a double
Interactables.Interactables namespace. Whilst this looks a bit strange
it at least buckets the code at the correct level.