feat(Interactables): provide helper methods for common tasks
The InteractableFacade now has some common helper methods for
tasks like enabling/disabling the touch and grab configurations
in case of toggling the interactable behaviour.

It is also possible to get access to common sub components from
the facade such as the Rigidbody, MeshContainer, a collection
of MeshRenderers and a collection of Colliders.

The Interactable.TouchReceiver prefab has also been updated to
contain additional useful references.