fix(Interactions): force action grab setup to occur immediately
There is an issue where instantiating an Interactable prefab and
then immediately calling an Interactor.Grab on it will cause a
null exception because for some reason Unity hasn't fully set up
the instantiated GameObject by the time the event process has run
and therefore the `action.GrabSetUp` won't be set until the `action`
GameObject is enabled in the scene.

The simplest solution to this is to force set the `action.GrabSetUp`
immediately and then still do the wait to set it when the GameObject
is enabled and then this way it will exist to do the preliminary
setup and any additional setup that requires an active action will
still occur.