fix(Interactions): allow toggle grab to work when no longer touching
The Toggle Grab mechanism would not work if the first toggle would grab
and then the interactor would stop touching the interactable as the
second toggle would no longer release the grab.

This was due to the grab publishers were always populated based on
what was being touched and if the interactor stopped touching the
grabbed interactable then the start grab publisher would not have
anyway of knowing which interactable to publish to.

The latest version of Zinnia (1.22.0) has a mechanism for registering
successful consumers of published events and this is now being used
to keep a copy of all consumers that have successfully been published
to by the interactor and then even if the interactor stops touching
the interactable, the interactor still knows which consumers it can
publish to (the grabbed consumers) because of this registry.

The ActionPublisher has also been updated to utilise this new Zinnia
consumer registry to ensure the same concept of a grabbed interactable
can still pass through 3rd party button data even if it is not being
touched by additionally registering to the interactor start grab
publisher.