feat(Interactions): allow follow as standard secondary action
The secondary action can now also be set to follow, which allows the
primary action to be set to none so the primary action does nothing
whereas the secondary action will then grab the item to mimic two
hand grabbing.

The None action now has its own null action rather than just using the
base action so it can be easily identified as a nothing action.