feat(Interactions): add ability to choose valid interaction types
The new ValidInteractionTypes enum property makes it possible to
pre-determine what interaction types the Interactable will respond to.

Note: A grab can still occur if the Interactable does not respond to
a touch because the touch on the Interactor side is still valid. This
is useful for disabling touch events but still wanting to grab.

An UngrabAll method has been added too that is just a shortcut to
`Ungrab(0)` for easier understanding in the facade.