fix(Interactions): prevent attach point with precision grab
The Grab Attach Point on the Interactor is no longer used when an
Interactable is set to precision grab because if the Grab Attach Point
has a Transform offset then it means the precision grab is offset too.

Instead, a new Precision Attach Point has been added to the Interactor
that has a Vector3.zero position so this is used as the follow Source
for precision grab so the offset is always zero.

It is still possible to offset this new Precision Attach Point if
required, but the effects will be the precision grab is also offset.