feat(prefabs): update ControlDirection to implement follow TargetOffset
The latest version of the DirectionModifier now allows for a
TargetOffset to be provided, which can be provided from any Follow
action as they also consider a TargetOffset. The first found
TargetOffset of a complimentary follow action on an interactable will
be used as the TargetOffset on the ControlDirection Directionmodifier.