feat(Interactors): add ability to ignore colliders based on tag
A recent update to the Collision Tracker means it is now possible
to ignore specific colliders based on a rule. This is now being used
in the Interactor to ignore any collider with the
IgnoreInteractorOnColliderTag component on it. This can now be added
to any collider on an interactable to make the interactor ignore the
collider on the interactable.

This can be useful for example to ignore the blade of a sword so it
can only be grabbed by the handle.