feat(Interactor): add ability to simulate a touch with no collision
The new SimulateTouch and SimulateUntouch method on the
InteractorFacade allow for the Interactor to effectively touch an
Interactable without needing to physically collide with it in the
spatial world.

The Grab() method has also been updated so it uses this SimulateTouch
before grabbing so all of the correct events are triggered.