feat(Interactables): implement interactable grabber
The InteractableGrabber component actually existed in the
PointerInteractors package but it is far more general purpose than
just being used for pointer interactions. So the component has been
included in this repo where it can serve a wider use and it can then
be used to supersede the one in the PointerInteractors package.