fix(Touching): only call touch consumers if touch has changed
The Slicer component now provides a mechanism to determine if the
sliced list contents have changed so this makes it more efficient
when needing to call the interactable collision consumers.

Previously, they were being called every OnStay even if the current
touched object hadn't changed and this was very inefficeint and could
lead to a drop in framerate if touching many objects.