feat(Interactions): multiply velocity on interactor and interactable
The Interactor and Interactable now has a VelocityMultiplier on them
so the throw velocity can be multiplied on either the Interactor
or the Interactable (or both).

The Interactable Editor has also been tidied up a bit with some
horizontal lines added and a bit more indentation to make it easier
to follow the flow of data.

A number of enum properties have also had set methods added for them
as well.