feat(ActionPublisher): remove need for string identifier
The ActionPublisher no longer needs a string for the publisher
identifier as this was just being used intenally in the rules. But this
has now been replaced with a List Contains Rule and it just uses the
actual ActionPublishers that are linked onto the ActionReceiver to make
the internal rule work.

The FirstTouched logic was also broken so that has been fixed as part
of this commit.