feat(Action): provide rule based orientation handle option
The GrabInteractableFollowAction and the GrabAction.Follow prefab have
now been updated to also allow for the orientation handle to be set via
a rule.

This makes it possible to determine orientation handles by a rule at
runtime, so an interactable can be instantiated and does not need any
custom code to hook up the GameObject Relation.