feat(Utility): add streamline way to generate interactables via code
The new InteractableCreator prefab makes it much easier to create an
Interactable Object via code by simply calling the `Convert` method
passing in the GameObject to convert (and wrap) to an Interactable.

The core generation code has now been moved into a factory that the
Interactable Creator Editor Window now also uses. The ability to
set the primary and secondary actions has also been moved to the
InteractableConfigurator as it already holds all of the relevant
prefab references to be able to set an action.