fix(InteractableConfigurator): reset instantiated action orientation
When the interactable action is instantiated it was not having the
position, rotation and scale reset so if the original converted
object was not at the zero orientation then the action was being
offset.

This fix ensures the position, rotation and scale is reset to the
identity when the action is childed to the interactable.