fix(InteractableConfigurator): set action instantiation correctly
There was an issue when instantiating a new action it was not possible
to update the position, rotation or scale of any child object in the
action GameObject. This has now been fixed by making the action object
match the world orientation/scale of the interactable facade so when
it is childed it resets to the correct zero origin.

The Instantiate method is also always called if the app is playing
and not the PrefabUtility method as this is only now used at edit
time in the editor.