feat(Editor): add sub options for action types
A new sub option has been added for the `None` action and the `Scale`
action that bubbles up useful internal settings.

The primary/secondary action drop downs also now have a button next
to them that pings the action gameobject within the interactable
hierarchy to make it easier to navigate to.

Also, the entire editor script has been refactored to make it tidier
and more concise to read.