feat(Interactions): add IsVisible property to interactable
The IsVisible property on the getter will return whether the
interactable object is in its `visible` state which is the renderers
are active and the colliders working.

When the IsVisible property is set to `false` then all of the renderers
are disabled and all of the colliders are set to trigger colliders to
mimmic the GameObject becoming invisible but still active within the
scene.