feat(Grab): add option to attempt swap grab options
A new option that provides the ability to attempt to swap the secondary
grab action to the primary grab action when the secondary grab is
still active when the primary action ungrabs.

This also fixes an issue where using Primary:Follow and
Secondary:ControlDirection causes the rotation upon release of the
secondary interactor to not stay in the correct rotation when
precision grab is used because the precision grab container rotation
wasn't being updated with the control direction pivot.