fix(Grab): remove hard requirement for rigidbody
There is no definitive requirement for a rigidbody on an Interactable
but the previous changes that prepared rigidbody for kinematic changes
has meant that a Rigidbody is required to set up the state to prepare
for any kinematic change.

This fix just makes sure these state preparations are only done if a
rigidbody is provided, so it no longer needs a rigidbody if not
required.