fix(Configurator): catch error on destroy nested prefabs
Older versions of Unity cannot destroy a nested prefab, so as the
Controllables for example nest the Interactable prefab, if the
Interactable prefab is attempted to change the grab action then
this tries to delete the nested grab action but will raise an
InvalidOperationException because this nested prefab cannot be
deleted.

The fix for this is to just disable the hard wired prefab and then
create a new temporary prefab that can be continually updated.