feat(Facade): control rigidbody kinematic state from IsVisible
The IsVisible property also now controls the rigidbody kinematic state
otherwise an invisible object will just fall through all other
colliders in the scene.

A helper component of InteractableVisibilityStatusTag has also been
added to the top level of the Interactable prefab that is enabled
when the Interactable is visible and disabled when the Interactable
is hidden.

This can be used in rules to ignore hidden Interactables if needed.