fix(Prefabs): disallow collision consumers for invalid interactor
The DisallowedGrabInteractors rule was causing an issue when using
precision grab due to the collision consumer still being published
for an invalid interactor which was causing the precision point
container to be created for an invalid interactor. This meant that
any other valid interactor was not becoming a precision point grab
as the detached precision point for the invalid interactor still
existed.

The solution is to modify the GrabReceiver prefab so it turns off
the OutputActiveCollisionConsumer logic if the interactor is disallowed
and then the precision point cannot be created prior to a grab event.