feat(Facade): allow multiple rules for disallowed touch/grab types
The disallowed touch and grab interactors used to only work on a
ListContainsRule and this was hard coded into the facade editor.
This meant that no other rule could be replaced as it would throw
an error due to the custom editor.

This feature introduces a new AnyRule at the heart of both the
touch/grab disallowed logic with the first rule being the existing
ListContainsRule, so all of the existing logic still works and
that ListContainsRule must be at the zero index of the AnyRule list.

But now the AnyRule can also contain any other rule that is required
to disallow touch or grab interactors.